[brain tests] Burst nose from the main file
